Extract the network tooltip contents into its own file
